pulling my hair out

+++ include/simulation/physics.h
@@ -1,13 +1,13 @@
 #ifndef PHYSICS_H
 #define PHYSICS_H
 
-float dP(float M_r, float r, float rho);
-float dT(float L_r, float r, float T, float kappa, float rho);
-float rho(float P, float T);
-float epsilon(float T, float rho);
-float kappa(float T, float rho);
-float dM(float r, float rho);
-float dL(float dM, float epsilon);
+double dP(float M_r, float r, float rho);
+double dT(float L_r, float r, float T, float kappa, float rho);
+double rho(float P, float T);
+double epsilon(float T, float rho);
+double kappa(float T, float rho);
+double dM(float r, float rho);
+double dL(double dM, double epsilon);
